groundwork for autobuilding
